rename: 파일명 index -> 컴포넌트명으로 변경
