+++ src/components/pages/Layout/Layout.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { StLayout } from './styled';
 import { Outlet } from 'react-router-dom';
-import Footer from './Footer';
-import Header from './Header';
+import Footer from './footer/Footer';
+import Header from './header/Header';
 
 const Layout = ({children}) => {
   return (
