+++ src/components/shared/Input/inputSearch/InputSearch.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Input from '../index'
-import InputBox from '../../InputBox'
+import InputBox from '../../inputBox/InputBox'
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSearch, faCircleXmark } from "@fortawesome/free-solid-svg-icons";
